Hacks? Fixes? Getting performance ready!
