+++ README.md
@@ -9,12 +9,12 @@ shift+1 to shift+4 will give you preset screens. (currently for the madtixx dj s
 
 each line of a qwerty keyboard will control one of the layers: q-p is for the background bank, 1-8 for the background modifier, a-; for the overlay bank, z-, for the overlay modifier
 
-| Level | Effect+ | Effect- | Var+ | Var- | Time Mode+ | Time Mode- | Time+ | Time- | Palette+ | Palette- |
-| ---: | :---: | :---: | :---: | :---: | :---: | :---: | :---: | :---: | :---: | :---: |
-| Back Modifier | 1 | 2 | 3 | 4 | 5 | 6 | 7 | 8 |
-| Back Bank Effect | Q | W | E | R | T | Y | U | I | O | P |
-| Front Bank Effect | A | S | D | F | G | H | J | K | L | ; |
-| Front Modifier | Z | X | C | V | B | N | M | < |
+| Level | Effect- | Effect+ | Var- | Var+ | Time Mode- | Time Mode+ | Time- | Time+ | Palette+ | Palette- | Reset |
+| ---: | :---: | :---: | :---: | :---: | :---: | :---: | :---: | :---: | :---: | :---: | :---: |
+| Back Modifier | 1 | 2 | 3 | 4 | 5 | 6 | 7 | 8 | ALT+1 |
+| Back Bank Effect | Q | W | E | R | T | Y | U | I | O | P | ALT+Q |
+| Front Bank Effect | A | S | D | F | G | H | J | K | L | ; | ALT+A |
+| Front Modifier | Z | X | C | V | B | N | M | < | ALT+Z |
 
 in each line, controls are grouped in 2 keys. 
   the first two keys (ie, q and w) will rotate through the various options. 
@@ -23,6 +23,20 @@ in each line, controls are grouped in 2 keys.
   (u and i) divide the time variable, and you can go negative as well
   for the banks, the last key pair (o and p) change the palettes
 	
+- SHIFT + BACKSPACE - Dump state to Trace
+- SHIFT + (0-9) - Preconfigured shortcuts
+- SHIFT + Home - FFT increase
+- SHIFT + End - FFT decrease
+- Home - Loudness increase
+- End - Loudness decrease
+- Insert - cls each frame toggle (effect)
+- Delete - cls each frame toggle (overlay)
+- SHIFT + Insert - Stutter (switching cls) effect
+- SHIFT + Delete - Stutter (switching cls) overlay
+- Page up - effect <-> modifier order
+- Page down - overlay <-> modifier order
+- TAB - toggle debug on screen
+
 pressing space 4 times will change the beat timing, so press it to beat match to the kickdrum of some EDM
 
 insert and delete control clearscreen for the background and overlay layers (so you can see the modifiers)
