+++ code/globals.lua
@@ -1,5 +1,5 @@
 m=math
 sin,cos,max,min=m.sin,m.cos,m.max,m.min
 abs,pi,rand,randseed,exp=m.abs,m.pi,m.random,m.randomseed,m.exp
-sqrt,atan2=m.sqrt,m.atan2
+sqrt,atan2,floor=m.sqrt,m.atan2,m.floor
 tau=pi*2
