+++ debug/fakefft.lua
@@ -1,6 +1,7 @@
 if fft == nil then
     -- Not great =^D
     function fft(v)
-        return (0.5+(sin(v)^2)*.5)/v
+		local t = (time()%1000)/1000
+        return 0.7+((sin(v+t*.001)^2)*.3)
     end
 end
