+++ palette/grey_scale.lua
@@ -6,7 +6,7 @@ return {
 	id="grey_scale",
 	boot=function()
 		for i=0,47 do
-			palette[i]=i*5.6
+			palette[i]=i*5.4
 		end
 	end,
 	get=function(y, t)
